+++ src/com/nastel/jkool/tnt4j/limiter/LimiterFactoryImpl.java
@@ -24,12 +24,12 @@ package com.nastel.jkool.tnt4j.limiter;
 public class LimiterFactoryImpl implements LimiterFactory {
 
 	@Override
-    public Limiter newLimiter(int maxMps, int maxBps) {
+    public Limiter newLimiter(double maxMps, double maxBps) {
 	    return new LimiterImpl(maxMps, maxBps, true);
     }
 	
 	@Override
-    public Limiter newLimiter(int maxMps, int maxBps, boolean enabled) {
+    public Limiter newLimiter(double maxMps, double maxBps, boolean enabled) {
 	    return new LimiterImpl(maxMps, maxBps, enabled);
     }
 }
